Refactor table out
