+++ src/referrals-queue/referrals-queue.css
@@ -8,18 +8,8 @@
   max-width: 960px; /* 2 */
 }
 
-.table {
-  border: 2px solid grey;
-  width: 100%;
-}
-
 .tableContainer {
   height: 500px;
   overflow-x: hidden;
   overflow-x: auto;
-}
-
-.tr {
-  border: 1px solid grey;
-  line-height: 2rem;
 }
