+++ src/table/table.css
@@ -1,0 +1,9 @@
+.table {
+  border: 2px solid grey;
+  width: 100%;
+}
+
+.tr {
+  border: 1px solid grey;
+  line-height: 2rem;
+}
